Improve icon organization and related file docs
