+++ declarations.d.ts
@@ -1,3 +1,6 @@
+/**
+ * Adds a type to SVG imports on mobile.
+ */
 declare module '*.svg' {
 	import { SvgProps } from 'react-native-svg';
