+++ src/elements/Icon/index.ts
@@ -2,3 +2,4 @@ import Icon from './Icon';
 
 export { Icon };
 export * from './icons';
+export * from './Icon.types';
